helper/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.8
pycryptodome==3.9.4